// libs/feature/manager/offer/request/list/component/src/lib/types/offer-request-columns.enum.ts
export enum OfferRequestColumnEnum {
    Date = 'date',
    Status = 'status',
    Offer = 'offer',
    RequestedBy = 'requested_by',
    Affiliate = 'affiliate',
    Fraud = 'fraud',
    Questions = 'questions'
}

export enum OfferRequestColumnSortEnum {
    Date = 'created',
    Offer = 'offer_id',
    Affiliate = 'affiliate_id'
}